refactor(auth): セッション購読を SessionProvider に集約し get-session を1回に
better-auth の useSession は購読ごとに初回 get-session を投げ、同時フェッチは
dedupe されない。CursorLayer・各ページ・ProtectedRoute が個別に useSession して
いたため、ロード時に get-session が複数同時発火していた。

SessionProvider を唯一の購読者としてルート最上位 (router.tsx) に置き、結果を
context で配布する。各コンポーネントは @/lib/session の useSession を使うだけで、
セッション取得の作法を一切意識しなくてよい。返り値の形は better-auth と同一なので
呼び出し側の変更は import 元の差し替えのみ。
